move nomralizeConfig to sdk, fix inplace sorts
