+++ typescript/sdk/src/ism/EvmIsmModule.ts
@@ -26,7 +26,6 @@ import {
   assert,
   deepEquals,
   eqAddress,
-  normalizeConfig,
   objFilter,
   rootLogger,
 } from '@hyperlane-xyz/utils';
@@ -46,6 +45,7 @@ import { ContractVerifier } from '../deploy/verify/ContractVerifier.js';
 import { MultiProvider } from '../providers/MultiProvider.js';
 import { AnnotatedEV5Transaction } from '../providers/ProviderType.js';
 import { ChainName, ChainNameOrId } from '../types.js';
+import { normalizeConfig } from '../utils/ism.js';
 import { findMatchingLogEvents } from '../utils/logUtils.js';
 
 import { EvmIsmReader } from './EvmIsmReader.js';
